+++ src/config.py
@@ -67,32 +67,34 @@ ThinkingEffortLevel = Literal[
 StructuredOutputMode = Literal["json_schema", "json_object"]
 
 
-PROVIDER_TIMEOUT_ERROR = "provider_params.timeout must be a positive number of seconds"
+PROVIDER_TIMEOUT_ERROR_TEXT = (
+    "provider_params.timeout must be a positive number of seconds"
+)
 
 
 def coerce_provider_timeout(value: Any) -> float:
     """Coerce a `provider_params.timeout` value to positive, finite seconds.
 
     Canonical implementation shared by config-load validation (here) and
-    per-request validation (`src.llm.backend.request_timeout_from_extra_params`,
+    per-request validation (`src.llm.request_builder.request_timeout_from_extra_params`,
     which translates the ValueError into a ValidationException). Lives in
     config.py because src.exceptions imports src.config, so config validators
     cannot raise Honcho exception types.
     """
     if isinstance(value, bool):
-        raise ValueError(PROVIDER_TIMEOUT_ERROR)
+        raise ValueError(PROVIDER_TIMEOUT_ERROR_TEXT)
     if isinstance(value, int | float):
         timeout = float(value)
     elif isinstance(value, str):
         try:
             timeout = float(value.strip())
         except ValueError as exc:
-            raise ValueError(PROVIDER_TIMEOUT_ERROR) from exc
+            raise ValueError(PROVIDER_TIMEOUT_ERROR_TEXT) from exc
     else:
-        raise ValueError(PROVIDER_TIMEOUT_ERROR)
+        raise ValueError(PROVIDER_TIMEOUT_ERROR_TEXT)
 
     if not math.isfinite(timeout) or timeout <= 0:
-        raise ValueError(PROVIDER_TIMEOUT_ERROR)
+        raise ValueError(PROVIDER_TIMEOUT_ERROR_TEXT)
     return timeout
