+++ src/llm/request_builder.py
@@ -11,14 +11,13 @@ from typing import Any, cast
 
 from pydantic import BaseModel
 
-from src.config import ModelConfig, PromptCachePolicy
+from src.config import ModelConfig, PromptCachePolicy, coerce_provider_timeout
 from src.exceptions import ValidationException
 
 from .backend import (
     CompletionResult,
     ProviderBackend,
     StreamChunk,
-    request_timeout_from_extra_params,
 )
 
 # Operator escape-hatch keys recognized inside ModelConfig.provider_params.
@@ -104,12 +103,43 @@ def build_config_extra_params(config: ModelConfig) -> dict[str, Any]:
     return extra_params
 
 
+def request_timeout_from_extra_params(
+    extra_params: dict[str, Any] | None,
+) -> float | None:
+    """Return a validated per-request provider timeout from extra params.
+
+    Config-sourced timeouts are already validated and normalized at config
+    load (`coerce_provider_timeout` in src.config); this guards extra_params
+    passed programmatically at call time.
+    """
+    if not extra_params or "timeout" not in extra_params:
+        return None
+
+    try:
+        return coerce_provider_timeout(extra_params["timeout"])
+    except ValueError as exc:
+        raise ValidationException(str(exc)) from exc
+
+
+def _strip_none_params(
+    params: dict[str, Any],
+    keys: tuple[str, ...],
+) -> dict[str, Any]:
+    """Remove specified keys from extra params when their values are None."""
+    return {k: v for k, v in params.items() if not (k in keys and v is None)}
+
+
 def _normalize_extra_params(extra_params: dict[str, Any]) -> dict[str, Any]:
-    """Normalize shared extra params before they reach provider backends."""
-    timeout = request_timeout_from_extra_params(extra_params)
-    if timeout is None:
-        return extra_params
-    return {**extra_params, "timeout": timeout}
+    """Normalize and clean shared extra params before they reach backends.
+
+    Centralizes per-key coercion and null-stripping so new keys are added
+    here rather than spawning one-off normalizers.
+    """
+    result = dict(extra_params)
+    timeout = request_timeout_from_extra_params(result)
+    if timeout is not None:
+        result["timeout"] = timeout
+    return _strip_none_params(result, ("timeout",))
 
 
 async def execute_completion(
